+++ authlete/dto/token_request.py
@@ -1,5 +1,5 @@
 #
-# Copyright (C) 2019 Authlete, Inc.
+# Copyright (C) 2019-2020 Authlete, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -27,7 +27,10 @@ class TokenRequest(Jsonable):
             'clientSecret':          str,
             'clientCertificate':     str,
             'clientCertificatePath': str,      # list of str
-            'properties':            Property  # list of Property
+            'properties':            Property, # list of Property
+            'dpop':                  str,
+            'htm':                   str,
+            'htu':                   str,
         }
 
         super().__init__(nameAndValues, nameAndTypes)