+++ CHANGES.md
@@ -75,6 +75,11 @@ CHANGES
 - `TokenIssueResponse` class
     * Added `accessTokenResources`.
 
+- `TokenRequest` class
+    * Added `dpop`.
+    * Added `htm`.
+    * Added `htu`.
+
 
 1.0.2 (2019-08-09)
 ------------------